+++ servers/audio/audio_driver_dummy.cpp
@@ -46,7 +46,7 @@ Error AudioDriverDummy::init() {
 	int latency = GLOBAL_GET("audio/driver/output_latency");
 	buffer_frames = closest_power_of_2(latency * mix_rate / 1000);
 
-	samples_in = memnew_arr(int32_t, buffer_frames * channels);
+	samples_in = memnew_arr(int32_t, (size_t)buffer_frames * channels);
 
 	thread.start(AudioDriverDummy::thread_func, this);
